Add files via upload
The remaining html pages and the images (make sure the images are inside an image folder, or our index html won't have access to them and they won't upload on our live server). Also, I included the bootstrap html code - incase Lachlan asks for the template or we need the code for future pages.